requirements.txt
appnope==0.1.0
attrs==19.3.0
backcall==0.2.0
Brotli==1.0.7
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
click-plugins==1.1.1
cligj==0.5.0
cssmin==0.2.0
dash==1.12.0
dash-core-components==1.10.0
dash-html-components==1.0.3
dash-renderer==1.4.1
dash-table==4.7.0
decorator==4.4.2
Fiona==1.8.13.post1
Flask==1.1.2
Flask-Assets==2.0
Flask-Compress==1.5.0
Flask-SQLAlchemy==2.4.3
Flask-WTF==0.14.3
future==0.18.2
geographiclib==1.50
geopandas==0.7.0
geopy==1.22.0
idna==2.9
ipykernel==5.3.0
ipython==7.15.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.17.0
Jinja2==2.11.2
jsmin==2.2.2
jupyter-client==6.1.3
jupyter-core==4.6.3
lesscpy==0.14.0
MarkupSafe==1.1.1
munch==2.5.0
mysql-connector==2.2.9
nominatim==0.1
numpy==1.18.4
pandas==1.0.4
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
plotly==4.8.1
ply==3.11
prompt-toolkit==3.0.5
ptyprocess==0.6.0
Pygments==2.6.1
pyproj==2.6.1.post1
pyScss==1.3.7
python-dateutil==2.8.1
python-dotenv==0.13.0
pytz==2020.1
pyzmq==19.0.1
requests==2.23.0
retrying==1.3.3
Shapely==1.7.0
six==1.15.0
SQLAlchemy==1.3.17
tornado==6.0.4
traitlets==4.3.3
urllib3==1.25.9
wcwidth==0.2.4
webassets==2.0
Werkzeug==1.0.1
WTForms==2.3.1
